Fixed warnings

+++ tests/Delay/FixedDelayLineFilter.cpp
@@ -31,7 +31,7 @@ BOOST_AUTO_TEST_CASE( FixedDelayLineFilter_line100_delay_100_test )
 BOOST_AUTO_TEST_CASE( FixedDelayLineFilter_sinus_line100_delay50_test )
 {
   boost::scoped_array<float> data(new float[PROCESSSIZE]);
-  for(int64_t i = 0; i < PROCESSSIZE; ++i)
+  for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = std::sin(2 * boost::math::constants::pi<float>() * (i+1.)/48000 * 1000);
   }
@@ -55,12 +55,12 @@ BOOST_AUTO_TEST_CASE( FixedDelayLineFilter_sinus_line100_delay50_test )
   output.process(51);
   output.process(PROCESSSIZE - 1 - 49 -51);
   
-  for(int64_t i = 0; i < 50; ++i)
+  for(ptrdiff_t i = 0; i < 50; ++i)
   {
     BOOST_REQUIRE_EQUAL(0, outdata[i]);
   }
   
-  for(int64_t i = 50; i < PROCESSSIZE; ++i)
+  for(ptrdiff_t i = 50; i < PROCESSSIZE; ++i)
   {
     BOOST_REQUIRE_EQUAL(data[i - 50], outdata[i]);
   }
